Accept and evaluate 'needManualCorrection' attribute on ClozeTask gaps.
This is an IALT-specific issue. Here, a gap in a cloze task can be
corrected automatically, if none of the correct answers match the answer
given by the examinee.

This change accepts 'needManualCorrection' attribute on gaps in a
ClozeTask. It will also accept this flag in the <cloze-config> element,
but won't take this flag into account (it's just there for validation).